accounting_app/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
whitenoise==6.6.0
dj-database-url==2.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
django-htmx==1.17.2
reportlab==4.0.8
openpyxl==3.1.2
psycopg2-binary==2.9.9
Faker==20.1.0